+++ crates/primitives/Cargo.toml
@@ -15,6 +15,7 @@ alloy-rlp = { workspace = true, features = ["derive"] }
 alloy-primitives = { workspace = true, features = ["rlp"] }
 
 # External
+async-trait = "0.1.77"
 hashbrown = "0.14.3"
 unsigned-varint = "0.8.0"
 spin = { version = "0.9.8", features = ["mutex"] }
@@ -37,3 +38,4 @@ serde_json = { version = "1.0.68", default-features = false }
 [features]
 default = ["serde"]
 serde = ["dep:serde"]
+test-utils = []
